+++ runtime/observatory/lib/src/service/object.dart
@@ -2204,6 +2204,9 @@ class ServiceEvent extends ServiceObject {
   static const kConnectionClosed = 'ConnectionClosed';
   static const kLogging = 'Logging';
   static const kExtension = 'Extension';
+  static const kTimelineEvents = 'TimelineEvents';
+  static const kTimelineStreamSubscriptionsUpdate =
+      'TimelineStreamSubscriptionsUpdate';
   static const kServiceRegistered = 'ServiceRegistered';
   static const kServiceUnregistered = 'ServiceUnregistered';
   static const kDartDevelopmentServiceConnected =
@@ -2237,6 +2240,7 @@ class ServiceEvent extends ServiceObject {
   String extensionKind;
   Map extensionData;
   List timelineEvents;
+  List<String> updatedStreams;
   String spawnToken;
   String spawnError;
   String editor;
@@ -2321,6 +2325,9 @@ class ServiceEvent extends ServiceObject {
     }
     if (map['timelineEvents'] != null) {
       timelineEvents = map['timelineEvents'];
+    }
+    if (map['updatedStreams'] != null) {
+      updatedStreams = map['updatedStreams'].cast<String>();
     }
     if (map['spawnToken'] != null) {
       spawnToken = map['spawnToken'];