+++ pkg/vm_service/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## 4.0.3 
+## 4.0.4
+- Update to version `3.34.0` of the spec.
+- Fixed issue where `TimelineEvents` was not a valid service event kind.
+- Added `TimelineStreamSubscriptionsUpdate` event, which is broadcast when
+  `setVMTimelineFlags` is used to change the set of currently recording timeline
+  streams.
+
+## 4.0.3
 - Update to version `3.33.0` of the spec.
 - Add static error code constants to `RPCError`.
 - Update the toString() method or `RPCError` and add a toMap() method.